# server/swagger.yaml
openapi: 3.0.3
info:
  title: MyLoadTest Datapool API
  description: The MyLoadTest Datapool is a centralised, persistent data store. ...TODO
  version: 0.0.2
  contact:
    name: MyLoadTest
    url: https://www.myloadtest.com
servers:
  - url: http://localhost:3000
    description: Self-hosted datapool
...
components:
  schemas:
    Counter:
      description: a counter value (must be a positive whole number)
      type: integer
      minimum: 0
      example: 99
...
    QueueItem:
      description: Any value may be added to the queue
      type: array
      items:
        anyOf:
          - type: string
          - type: number
          - type: object
      example:
        - orderNum: '0001'
          total: '99.50'
          postage: '10.50'
paths:
...
  /counter/{name}:
    post:
      summary: Initialise a counter
      description: Initialise a counter with a custom starting value. If you just need a counter that starts
        at 1, you can call ```GET /counter/{name}``` directly.
      tags: [counter]
      parameters:
        - in: path
          name: name
          description: The name of the counter
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Counter'
      responses:
        201:
          description: An empty response body is returned when the counter is initialised or reset.
        400:
          description: Bad request. Counter value must be a positive integer.
    get:
      summary: Get the value of a counter
      description: Get the value of a counter (counter names are case sensitive).
      tags: [counter]
      parameters:
        - in: path
          name: name
          description: The name of the counter
          required: true
          schema:
            type: string
      responses:
        200:
          description: Returns the counter value. Note that a 404 response is not returned if the counter does not exist. Instead, the counter is created and initialised to 1.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Counter'
        400:
          description: The counter name is invalid
    delete:
      summary: Remove or reset a counter
      description: Remove or reset a counter (no option to delete multiple counters at the same time)
      tags: [counter]
      parameters:
        - in: path
          name: name
          description: The name of the counter
          required: true
          schema:
            type: string
      responses:
        200:
          description: The counter was deleted successfully. Note that there is no point returning a 404 if the counter is not found, as it is the same outcome if it is deleted or never existed in the first place.
        400:
          description: The counter name is invalid
...
  /queue/{name}:
    post:
      summary: Save one or more items to the queue
      description: Save a list of items to the specifed queue. If the queue does not exist, it will be created.
      tags: [queue]
      parameters:
        - in: path
          name: name
          description: The name of the queue
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/QueueItem'
      responses:
        201:
          description: An empty response body is returned when the value is saved.
        400:
          description: Bad request. Saved value must be an array of items.
        413:
          description: Payload too large. The request body may be up to 100kb.
    get:
      summary: Get the next item from the queue
      description: Get the next item from the specified queue. To remove multiple items, append the ```?limit=x``` query string to the request. If the queue contains fewer items than specified by the limit, then only that number of items will be returned.
      tags: [queue]
      parameters:
        - in: path
          name: name
          description: The name of the queue
          required: true
          schema:
            type: string
        - in: query
          name: limit
          description: The maximum number of items to return (defaults to 1)
          required: false
          schema:
            type: integer
            minimum: 1
            example: 1
      responses:
        200:
          description: An array of JSON objects or primitive values
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QueueItem'
        400:
          description: The queue name is invalid
        404:
          description: The queue name was not found
    delete:
      summary: Clear the queue
      description: Delete the specified queue and all its queued items. There is no need to explicitly re-create a queue after clearing/deleting it as the first call to ```POST /queue/{name}``` will automatically create a queue with the specified name.
      tags: [queue]
      parameters:
        - in: path
          name: name
          description: The name of the queue
          required: true
          schema:
            type: string
      responses:
        200:
          description: The queue was deleted successfully. Note that there is no point returning a 404 if the queue is not found, as it is the same outcome if it is deleted or never existed in the first place.
        400:
          description: The queue name is invalid
        404:
          description: The queue name was not found
  /queue/{name}/depth:
    get:
      summary: Get the number of items on the queue
      description: Returns the number of items on the specified queue. The queue is empty when depth is zero.
      tags: [queue]
      parameters:
        - in: path
          name: name
          description: The name of the queue
          required: true
          schema:
            type: string
      responses:
        200:
          description: The item count for the queue
          content:
            application/json:
              schema:
                type: integer
        400:
          description: The queue name is invalid
        404:
          description: The queue name was not found
  /queue/{name}/peek:
    get:
      summary: Read the next item from the queue without removing it
      description: Read the next item from the specified queue, but do not remove the item from the queue.
      tags: [queue]
      parameters:
        - in: path
          name: name
          description: The name of the queue
          required: true
          schema:
            type: string
      responses:
        200:
          description: A JSON object or primitive value
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QueueItem'
        400:
          description: The queue name is invalid
        404:
          description: The queue name was not found
  /queue/{name}/export:
    get:
      summary: Read all items from the queue without removing them
      description: Read all the items from the specified queue, but do not remove any items from the queue.
      deprecated: true
      tags: [queue]
      parameters:
        - in: path
          name: name
          description: The name of the queue
          required: true
          schema:
            type: string
      responses:
        200:
          description: A JSON object or primitive value
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QueueItem'
        400:
          description: The queue name is invalid
        404:
          description: The queue name was not found
...